pyproject.toml
[project]
name = "sd3"
version = "0.1.0"
description = ""
authors = [
    {name = "armandoluz",email = "armandoari.288@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "jsonify (>=0.5,<0.6)",
    "pillow (>=11.1.0,<12.0.0)",
    "pyqt5 (>=5.15.11,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
